LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/script-element-json-src-expected.txt

CONSOLE MESSAGE: TypeError: 'application/json' is not a valid JavaScript MIME type.

PASS Test that <script> doesn't load when the src is JSON.
